Also glob images and descriptions
We want to glob these, because adding a new description should be as trivial as possible. So in this case, we just need to add a new json file and if necessary some images, run CMake again and package. We won't need to type anything in the CMake file. That's just busywork.