Remove some unused properties
You can't hover this because it's covered by a mouse area.
Always do depth indentation in this instantiation of the settings list.
Never do quality/user setting emphasis, since we don't care if the setting is overwritten by the user or not.
Remove stack levels properties because we don't do quality/user setting emphasis.