Store selected setting ID in CuraSettingsGuide and make data depend on it
This is now the leading authority on which setting is currently selected. Also the current setting data depends on it and will be obtained depending on that same property upon requesting it. There can no longer be any asynchronosity between the selected setting ID displayed in the sidebar and the title, and the actual setting data.

Fixes #1.