Add markers to explain what each position derivative is
It turns out though that the SVG renderer in Qt does not support the <marker> tag. So it looks like I need to use <use>. I wanted to commit this one so that I can confirm that it looks the same in Firefox if I need to stop at some point.