Implement basic Markdown parsing
Now we don't need those templates any more. It's going to get rid of a whole lot of QML. And this got rid of a whole lot of Python with all the image files and all the style parsing stuff.
Some of this is going to come back. I'm going to need some specific QML to put the images in a better place and to create a title and such.