Split images out of descriptions
Quite a lot of work, but it seems to be working now.
This finds all images from the descriptions. It puts the parts before and after the images and the images themselves in a list, and then renders the list with a repeater repeating a loader for either the text or the images. Quite complex, now, but very powerful.