Convert gradual_infill_step_height to Markdown, make corrections and re-do images
The original images didn't show the gradual infill effect very nicely since the model that was sliced had edges itself. Since infill was the only thing being shown, it looked like the gradual infill had a step there even though it was just the model being slimmer on the bottom than on the top.