Redo infill_before_walls gifs
The new GIFs have way fewer frames because they skip over some tiny travel moves and the model that they are printing is much smaller (which requires fewer infill lines). They also have the new infill colour, so that it'll be up to date for Cura 4.2 (insider information). They have higher quality dithering from Photoshop so they look nicer. But the file size is just an eighth of the original file. Better in every way.