Redo alternate_extra_perimeter.gif
The original GIF had like 40 frames where only 4 were meaningfully different. It seems to have been recorded with a screen recorder directly to GIF, instead of constructed frame by frame from screenshots.
The new GIF has the new colour for infill. It's higher resolution, less dithering issues due to compression, and only half the file size. The GIF was actually reduced to 16 colours, which is enough for this.