Typing fixes
One syntax error. Stupid.
Two times a dialogue that could've been set to None if the dialogue failed to load, in which case the plug-in would crash. It now saves one duplicate message less in the log so that's fine.
One case that should never happen because the plug-in always has an ID and a path by the time this function is called but anyway
And one other case where the _version can in theory be None but not really since the plug-in has been initialised long before we create the dialogue that needs the version number.