Convert wall_0_extruder_nr to Markdown, add information and redo images
I've redone all the wall_extruder_nr images in one go, in order to keep the model and camera in exactly the same position.
The images needed redoing because the original images were scaled down. I'm also not sure about the legality of that model and if I would be allowed to publish that picture.