Convert wall_line_width to Markdown, add information and redo images
Like the rest of the line width images, these images were just highlighting the walls rather than showing the effect of changing the line width.
This article is pretty good. It's an important setting after all. I'll copy it back and forth for the rest of the wall line width settings.