Only take essential parts from Mistune
Installing all the Git history can result in permission errors for some platforms. Previous releases I'd manually deleted them but this can be automated.
We need the mistune.py module itself as well as the license file. So this deletes the Git history, the docs, the tests, all of the CI files, the makefile, gitignore, change log and build files.