Use currentIndex of listview to mark selected item
This is necessary in order to use keyboard navigation with the ListView. The keyboard shortcuts are not working for me.
The keyboard shortcuts only work when the ListView has focus, right now. I don't know how to make that work. It works when you press Shift+Tab twice.
It's not yet highlighting the correct item when you click on a link in the article.