Reset currentIndex when loading a setting that's currently not visible
If a setting is not visible because its category is collapsed in the Settings Guide, the currentIndex would previously stay on the previous setting because there was no component owning up that the setting key was his. Now we set the currentIndex to -1 first, which makes the highlight disappear. Then if the setting is visible the highlight is moved to the correct location. Otherwise it stays invisible.
Because we also move the currentIndex on creation, if the user expands the category later the currentIndex is changed properly to the correct setting.