Automatically scale SVGs to correct size
So the actual intrinsic size of the SVG now doesn't matter any more. Small SVGs will properly be upscaled without any interpolation necessary.
This can also improve performance, depending on the size you're displaying the zoomed image at. The original zoomed image was meant to be downscaled for all but the largest screens, so for all but the largest screens the performance should be a little better when there's SVGs involved.