Add markdown rendering for inline images
This renderer corrects the source path for images to be relative to the articles. For now, this requires that all articles are in the main articles directory. They are now, but it won't work correctly any more when articles are subdivided into folders. We'll cross that bridge if we get there.
It sizes the images to be equal to the size of the tooltips for the settings list. This is to serve its current goal to be able to display images in the setting descriptions. That's not sustainable, I know. We'll see what we can do about that later.