Add article about flow rate compensation factor
This article contains the main explanation about this entire feature. Maybe it should be the first setting of the two. However the other setting is the one that turns it on or off right now.
Also correct some mistakes in the other article about when the compensation is applied. It's applied during the move following, not in between moves.