Make corrections and add information on adaptive layers threshold
I'm adding a more physical way of looking at this setting. The setting defines the horizontal distance between the layers' edges. That is what the tangent of the surface angle computes after all. Hopefully this is more intuitive than the abstract threshold used to be as well. We may want to rename the setting and reword its description as well though...