Add article about initial bottom layers
Love this new feature.
I can't make a screenshot of this in action yet though. I don't have the tools on my work laptop to create proper screenshots. I could make the screenshot there and edit it on my desktop but that's a hassle. I'll just wait until the beta for 4.4 is released.