Reduce colour depth of all screenshots
This reduces the colour depth of most images to a certain power of two such that the difference is very barely visible. A trained eye will see some differences, some banding or more pixely edges than before. But if you can't swap between the unreduced image and the reduced one it's extremely hard to see. And even for a trained eye it won't be bothersome since the story that the image wants to convey is still there.

Saves a surprising amount of data actually.