Require pre-adding all preference keys
This is necessary because the articles are not loaded in any particular order, but we still want to use the preference values during loading of the articles.
But you'll need to have the preferences in your code anyway since that is the only place where it can really be used.