Allow articles to be located in subdirectories
The list of articles is very long, making it hard to find the articles. So in order to group them up better I want to put them in subfolders. Previously it was only looking for articles in the one folder, but now it'll look for articles in subfolders as well.