Add more hardware causes of underextrusion
These are all of the common ones I can think of (leaving out the ones that are particular to my printer, such as dirt in the Bowden tube).
Next up, software mistakes. There are a LOT of those, so this will be a long article. I must attempt to keep it a bit brief. That'll be for another day though, I'm afraid. It's late.