Adjust the theme also in the build versions
For some reason, in the actual Cura builds, the main window's findChildren doesn't enumerate the main window rectangle. It just skips it and only enumerates the dialogues and pop-ups.
Instead we'll do this. Much more hacky, and it shows a duplicated entry in the themes list then. But it works.