Add article about the maximum park duration
Had to do a bit of research about this setting, as it was added during my vacation and without my knowledge, and the setting has no description to speak of in Cura. I actually have to dig through the firmware code to find this, since it's not documented ANYWHERE. Except here, now.