Don't store definitions inside a stack
This stack is unnecessary, since we only need to list the definitions themselves, not the stack.

The stack was also causing errors since it didn't have any extruders and thus Cura was viewing it as a pre-2.3 container stack that needed to be transformed. This was not implemented very nicely so it's still being executed on every stack to date. This was causing bug #13. This commit fixes #13.