Add diagram for the different dimensions of the print head
This one includes all of the sizes in one overview, so I won't have to make copies for the other settings.
It also includes a setting that is deleted in 4.5: Print Head Polygon. I've left it in for now, but once I can adjust the articles based on the Cura version I'll make a copy of the diagram with this one removed.