Revert Mistune 2.0 plug-in commits
This reverts commits:
* d29bb1f238dbbf90dd44b2083f24e019e6934d3a: Add regex matcher for new conditional syntax
* 3cda53354484ee906fe9c171fe4cc3fb345918c5: Add parser for conditional syntax
* 44148902d5c7e1f8a57b67e43bbe11c2d4e36b7f: Write renderer for conditional content

This was defining a plug-in for Mistune 2.0. However this approach didn't work for me at all. I think it was broken because the syntax matched on the same syntax as the built-in HTML blocks and inline HTML. So probably this was clashing. I'm going to use the renderer overrides instead for block HTML and inline HTML that comes built-in. That way I can also stay with the stable version of Mistune, which is an advantage to make it easy to build the plug-in.