Slightly simplify replacement pattern
We don't need a capturing group for all parts of the if statement here. We just want the whole thing. How exactly the whitespace in between worked doesn't matter, we can copy that along.
The group for the operator can't be removed since it makes the six options have greater precedence than the other regex operators.