Allow multiple conditions in one statement
You can now use a condition like 'cura_version > 4.2 and cura_version <= 4.5' to show certain text only in a specific range of version numbers, limited on both sides. I'm sure this will become more powerful later on if we ever get more variables.
The 'or' clause is not supported. You have to type the conditional text twice then. To implement an 'or' I'd need to have operator precedences and probably brackets or something, so I'm not getting into that yet until that bridge needs to be crossed.