Use only preprocessing to parse the conditional texts
This is in the end the only viable solution.
The previous solution would break if the conditional was spanning through previously existing Markdown syntax, because Mistune would give greater priority to some syntax elements. For instance, I had a conditional to render an extra bullet point at the end of a bullet list. This wouldn't work correctly, because the HTML tag would start in one item of the bullet list and end at the end of the next item. Mistune parses the bullet syntax first, and then sees that there is half an HTML comment at the end of the one bullet and half an HTML comment at the end of another, but doesn't link them together. It doesn't know what to do with that and so it just renders the HTML comment.
This is a bug in Mistune if you ask me, but since I'm currently using Mistune v1 I'm not in a position to file a bug report there unless I test it in v2.