Add avatar for contributors without Github account
I'm giving every contributor an image now, to streamline the contributor pages when there are more contributors and to give every contributor fair attribution (rather than some contributors having a picture there and some not). This also seems to be fair and private, in case people don't want their picture to be published there indefinitely.