Redesign the contributors page
I've decided on this format since it's more fair. This way, there is no more option for a flavour image to be put there. I'm also removing responsibility for maintaining people's privacy, if they have qualms with their picture being in this guide forever, as I'm using their avatar which they publish themselves.