Refer to Support Structure when talking about tree support
This is a bit problematic for the translations. The original Enable Tree Support article can't be linked to in newer Cura versions since it's no longer in the settings list nor in the Settings Guide's additional article list. The article isn't loaded any more. But the paragraph was clearly talking about two different settings you could choose from, which are now a single sentence. I did my best to keep the translation correct, only removing a part of one sentence and removing a hyperlink to the broken article.
It only changes for the web view and for Cura 4.7+. 4.6 and before should remain the same.