Move tree support settings to support category
This is where the settings are located now in Cura 4.7. For older versions of Cura, it doesn't matter that these files are in a different folder now; the Settings Guide finds the files regardless of which folder they are located in. The folders are purely for organisation so that we can more easily find the articles. It only really matters for the developers and for people who find the article on Github. For those, it will soon be more logical to look for them in the Support category since the latest Cura version will show them there.