Add test for links to articles that have already been translated
It's quite common that a translator would translate an article with a link to another article, and there is no translation yet for the other article. They are supposed to link to the untranslated version then. If then later the other article gets translated, there may already be dozens of links to the untranslated version and it's hard to track those down.
To help the translators, this test will identify any links that need to be updated to the translated versions.