Show setting guide in tooltips by default
I hope by now all of its bugs have been worked out. I haven't seen any problems myself nor heard of anyone else having those problems.
This is hopefully going to make it easier to discover the difference for a lot of people. Many users sort of assumed that the guide appears in the tooltips and were assuming that the old tooltips were the Settings Guide articles and were then a bit disappointed.