Better explanation of what happens when a mesh is not manifold
This is based on an explanation in Ultimaker/Cura#9133 which is an abstraction of the way in which the even-odd rule is applied to polygons. I've visualised that explanation which makes it way more intuitive than the previous way in which it was explained in this article, and transformed fieldOfView's explanation in a way that is a bit more complete in how this then applies to Cura's stitching method. It's not going into the deepest details but I think it's pretty good to understand this way.

The images needed to become a lot more complex because Qt's SVG renderer doesn't support gradients. The gradient there is absolutely crucial to visualise the area where it's ambiguous where the inside is and where the outside is. I've approximated the gradient with a sort of spiky pattern, which is not as nice to look at and certainly not as nice in code. But it's okay-ish.