Implement using Python scripts as model generators
Sometimes a model can't be generated by OpenSCAD, for instance if it's broken on purpose. In those cases we need to have a more powerful model generator, more free-form. A general-purpose language like Python can do this.
It's not the prettiest, but it does allow describing a 3D model in human-readable terms rather than just a long list of coordinates. It's just a bit more verbose than OpenSCAD, really.