Apply transformation only to correct mesh and apply per-object settings
We can now identify which node was just loaded, just by looking at the mesh data, since the mesh data is not copied or altered and will still exist.
Doing this, we can then apply the transformations only to the correct mesh, and also apply per-object settings to that mesh then.