Add instructions to reproduce ooze shield screenshot
This is using a LOT of the new features I just implemented. It's the first screenshot that really uses dual extrusion, so for that I needed to be able to load multiple models, to compile SCAD scripts with parameters, to give per-object settings to models and to change the layer view colour scheme.