Add screenshot instructions for circular prime tower
A bit of an awkward one since the setting doesn't exist any more, but the circular one should still be updated.
Also make sure that these are just taken with the default colour scheme. That's fine for these screenshots.
This completes the Dual category since the rest is inside an SVG. I'm not going to automate that one.