Add screenshot instructions for Make Overhang Printable settings
These two could be continued without further changes to the Screenshot Tool. For the next one I'll need to make more changes. Cura keeps coming up with novel ways in which the screenshots need to be made special! It's kind of annoying but also fun and rewarding since this screenshot tool becomes incredibly powerful this way.