Correct camera rotation when looking straight up or down
Previously the camera rotation depended on the previous location of the camera. It would strafe the camera to its correct position first, then rotate it to look at the target location. However this retained the up direction from its previous orientation. This rotated the picture when you ended up looking straight up or straight down. Now we set that to a canonical orientation. This canonical orientation is with the top of the picture pointing to positive Y when looking down or negative Y when looking up. That happens to be the correct orientation to reproduce all of the infill screenshots, and the correct orientation to match the infill density image.