Also allow layers to be repeated until length of lines
It'll now use as many as the maximum between the two. Layers is no longer the determining factor for how many frames you'll get. In fact, I suspect most animations will just get a single layer but a lot of different lines, because they show the animation of the nozzle moving.