Make copies of QImage data when screenshot is returned
This is necessary because QImage was just re-using the Numpy data without making a copy. The Numpy array then went out of scope and was deallocated. When making an animation, the next image was then allocated in the same memory location and written again. This overwrites the previous image, causing all images to be the same. Essentially, every image was referring to a Numpy array that was deallocated and this is undefined behaviour.
So make a copy, which forces Qt to make a copy of the Numpy array. This fixes animations.