Add screenshot instructions for infill patterns, density and offset
These now have to immediately update their images, since the grid infill pattern image is being reused in other articles now, and it would be best if the images are created in a consistent fashion in the same article, so the rest of the images in those articles need to be updated too. And the infill offset is reduced from 5 to 2. This is because the mesh is made smaller to be able to see the infill lines better, and at that smaller scale an offset of 2 is easier to see (an offset of 5 moves almost a compelete period of the pattern).