Update colour of alternative material in prime tower line width
The screenshot instructions - for consistency with other screenshots - use a yellow material rather than orange. The English text already had an updated colour in the description. I've updated the translations with the new colour and also update the image so that it's consistent with the text to how it will be generated eventually.