Add mappings for all of Cura's language codes
These are the most common ones that people could want to read the Settings Guide in.
I'm adding these in advance because some people seem to be experimenting with adding more languages to the Settings Guide. I want to help those people and make things easier for them. It'd be nice to make it work for them out of the box.