Fix frame optimisation of GIFs
So the frame optimisation of ImageMagick doesn't recognise that two colours are the same in different frames if they are colour-reduced. Apparently it compares the wrong things, i.e. the colour-reduced colour of a frame vs. the original colour of a different frame or something. In any case, you can't reduce the colours at the same time as optimising the frames. Optimising the frames was disabled anyway (I'd assume that FlexiGIF would do that, but no).
This drastically reduces the file size of GIFs if they have but small bits of movement. That allows for much greater framerates in these GIFs because greater framerate also quadratically reduces file size, to some extent. I'm doing 8FPS animations now for this next animation, which is about as good as it gets with the length of the lines in that one.