Improve reliability with more sleeps
We want to run this for extended periods of time, after all. This solved two issues for me:
- Sometimes Cura would crash with a segfault. With extended sleeps this doesn't happen any more.
- Sometimes screenshots would end up at lower resolution. Thankfully this doesn't happen any more either.

It also seems to help with how it sometimes wouldn't slice, even though the print is never seen. The extra sleeps helped I reckon.
There is one problem left, and that is that the entire process slows down towards the end due to having 500+ threads open. Even though the processes ended, the thread doesn't ever seem to end. This causes so much thrashing that the timers run out and it starts to try to slice before the models are properly loaded. You need to push the slice button manually then. I don't have time to fix that now.