Move top/bottom settings to their own category
This was moved in Cura 4.9, so I'll move them in this directory structure too. That way people can still easily find them online, even though the guide itself in Cura doesn't care in what folder the articles are located.
Downside is that existing links on the internet to the Github pages will break doing this. But in the long run it'll make things easier to find.
The tests will probably fail on this because links cross-referencing these settings are not updated. I'm doing that on purpose to make the changes easier to oversee. This way, this is purely about moving the files around and the files remain the same. In the next commit we'll have to fix the links.