Remove unused SettingPropertyProvider
This was giving a warning in the log, since activeMachineId doesn't exist any more. It was never filled with a container stack. That got me thinking though: Do we ever use that broken SettingPropertyProvider then? Turns out, no, we don't use it. Let's throw it away then. Fix a warning and improve performance slightly.