Re-apply description changes when global container stack changes
The definition of the global container has all of the setting descriptions. If the definition container is different, the setting descriptions should be changed again. So if the user changes their active printer, re-load the setting descriptions.
This fixes a bug where the articles weren't loaded in the tooltips after the user changed their printer.