Fix articles sometimes not rendering in tooltips
Comments were messing with Mistune. Mistune then failed completely, causing the Markdown to be shown in the tooltip. Due to rich text rendering, the Markdown was stripped of its newlines (Rich Text needs <br>) so the entire thing was unreadable and devoid of any layout or images.
Just remove the comments after the conditionals are processed. That works.