Add new combing mode in Cura 4.12
Since this new combing mode effectively replaces Not in Skin's usefulness completely, it requires some re-wording for the other one too. The rewording is not applicable before 4.12, so we must use conditional text for both.
I reworded the header sentence too to prevent having to use a conditional text for that too just because it said how many options there were. Conditional texts aren't really a problem, but it's a bit confusing for translators I think.