Fix broken inter-article links
Most of these were broken because there is no German translation for those articles yet. I could also have fixed those programmatically, by making the articles link automatically to the German if available and English article otherwise. However it also needs to work on Github when viewing the Markdown files, and there I have absolutely no influence on how the links are resolved. So we end up having to maintain actual proper links.

Tests are already in place to detect when one of these links still links to the English version if a German version is available, so it's minimal maintenance if the translations progress.

I suspect the bed_adhesion_problems.md article was machine-translated including the links, and the links were broken because of that.