Fix tests that finds broken links, and those links
The test was not finding broken links if there were multiple links in one paragraph and the first of those was broken, because of a regex matching greedily. By not matching greedily, the link text ends on the first closing square bracket on the line instead of the last, and so the link URL can be found. For a number of cases this was still a broken URL.

Shows how much we have to rely on those tests. I would never have found this, and it seems not a lot of people did (or those all were not motivated enough to report it).