Expand test to find broken links
It will now also find cases where a link got broken because a word is placed in between.
It might now be over-sensitive, but I really can't think of a case where you'd use square brackets other than the cases I've filtered out.
Hopefully this will help the translators, who are prone to this problem.