Fix link syntax to underextrusion article
However this then causes a different test to recognise it as a link, and the link is broken. It leads to a non-existing German translation for that article.
I'm leaving this in though (accepting the failed test) because pull request #69 will soon add the remaining German translations, including the Underextrusion article.