Create separate version of QML files for Cura 5
So many things broke now. And the layout is different too, so it'd be good to align the Settings Guide to the new layout. These are the major changes we need to look out for:
- No more RecolorImage
- Theme entries changed name
- Icons changed name
- onFoo properties deprecated
- New scrollbars
- The design of the sidebar changed

To support both Cura 4 and 5, we'll have to split off the code here.
Not everything is working correctly, but this is a good starting point.