Use Cura's style for the article scrollbars
For some reason, the article scrolling is slow and frictiony, while the settings list is smooth and slippery. And it's not just the scrollbar because before this change it was also the case. I know that the smooth scrolling is something from more recent Qt versions but I can't think of why it would be using an older implementation of ScrollView here.