Add article about Wall Ordering
This setting is basically the same as the old Outer Before Inner Walls setting. As an insider I know why it got changed: We wanted to add a third option originally, to print from inside out but printing the centremost wall last. That option got removed again though, opting to always print them last regardless. But we considered the drop-down to be easier to understand for the user than a checkbox setting. So it got changed to a drop-down.