Add article on Vertical Scaling Factor Shrinkage Compensation
As far as I could measure, there is no contraction effect on the vertical scale. There must be the ordinary shrinkage due to the plastic cooling off, but there were no internal stresses in the vertical direction in any of my prints. It really didn't shrink at all with PLA, and with PP it was too noisy to really make out if it shrank. So in this article I'm going by that the only relevant effect is the shrinkage due to cooling.
This makes the article a lot simpler.
There is also nothing relevant to explain about the positioning of models. I'm assuming that the reader understands that the model will scale down to the build plate (instead of hovering in mid-air after scaling or cutting the bottom off).